+++ yash-syntax/CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to `yash-syntax` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - ????-??-??
+
+### Changed
+
+- `syntax::HereDoc::content` redefined as `RefCell<Text>` (previously `Text`)
+- `syntax::RedirBody<H>::HereDoc::0` redefined as `Rc<H>` (previously `H`)
+    - `impl From<HereDoc> for RedirBody` replaced with `impl<T: Into<Rc<HereDoc>>> From<T> for RedirBody`
+- `parser::Lexer::here_doc_content` now taking a `&HereDoc` parameter and returning `Result<()>`
+- `parser::Parser::memorize_unread_here_doc` now taking an `Rc<HereDoc>` parameter
+
+### Removed
+
+- `parser::Parser::take_read_here_docs`
+- `parser::Fill`
+- `parser::MissingHereDoc`
+- Generic type parameters of AST types `RedirBody`, `Redir`, `SimpleCommand`, `ElifThen`, `CaseItem`, `CompoundCommand`, `FullCompoundCommand`, `FunctionDefinition`, `Command`, `Pipeline`, `AndOrList`, `Item`, `List` in the `syntax` module
+
 ## [0.2.0] - 2022-02-03
 
 Previously, source code attribution attached to ASTs was line-oriented. The